LRU Cache
This implements an LRU cache for often-serialized JSON objects. The LRU
cache's entries are refreshed upon a call to JSON.GET.

Cache invalidation happens whenever the given path (or any of its
children) is modified.

Entries that are small enough to not impact serialization on-the-fly are
not serialized.

Two 'private' commands, `JSON._CACHEINFO` and `JSON._CACHEINIT` are
provided for testing purposes.