Fix conflict on header strings.h
The POSIX header is need for the functions strcasecmp and strncasecmp. But a
header strings.h in RedisModuleSDK is prefered because of the -I specified.

Considering the way the library is used in the SDK, it seems a better practice
to specify that the header strings.h is in the rmutils directory and include
RedisModuleSDK straight.

Also, there is no need to have redismodule.h in the source as it is already
present in the SDK directory.