Abstract away the storage backend implementation.
Provides a separate trait for the storage backend implementation.

In the future, the storage backend trait may be removed from the
WhiteHolder trait and be used separately, allowing to change the backend
without changing the holder implementation.

Also removes the redundant `Box` usage in the traits in favour of "impl
Trait".