Correct the memory consumption for the custom parser.
Corrects the memory consumption calculations for the custom parser,
except that the Value::Object calculations aren't complete due to
custom implementations and other various factors (remains TODO).

Also adds a few more unit tests, which are a little less difficult
to work with and compare the values with the parsers.