Add a custom allocator trait and Redis implementation.
Allows to use a custom allocator specific to JSON values (can be used for
any other types too). The provided allocator trait is a drop-in
replacement for the unstable std::alloc::Allocator trait, which is prone
to change at this moment.

At the same time, this commit brings an auto-implementation of the
aforementioned trait for all other types, which implement the unstable
std::alloc::Allocator trait, to be used as a direct replacement, at
least for temporary purposes, until the standard trait becomes stable
ans safe to use.